+++ metatensor-core/tests/cpp/tensor.cpp
@@ -310,7 +310,11 @@ TEST_CASE("TensorMap") {
         ));
         tensor = TensorMap(Labels({"keys"}, {{0}}), std::move(blocks));
 
-        CHECK_THROWS_WITH(tensor.clone(), "external error: calling mts_array_t.create failed (status -1)");
+        CHECK_THROWS_WITH(tensor.clone(), "can not copy this!");
+
+        // reset the last error, this should free the exception and avoid memory leak
+        auto status = mts_set_last_error(nullptr, nullptr, nullptr, nullptr);
+        CHECK(status == MTS_SUCCESS);
     }
 
     SECTION("clone metadata") {
@@ -320,7 +324,11 @@ TEST_CASE("TensorMap") {
         CHECK(clone.keys() == tensor.keys());
 
         auto block = clone.block_by_id(0);
-        CHECK_THROWS_WITH(block.values(), "error in C++ callback: can not call `as_dlpack` for an EmtpyDataArray");
+        CHECK_THROWS_WITH(block.values(), "can not call `as_dlpack` for an EmtpyDataArray");
+
+        // reset the last error, this should free the exception and avoid memory leak
+        auto status = mts_set_last_error(nullptr, nullptr, nullptr, nullptr);
+        CHECK(status == MTS_SUCCESS);
     }
 }